BusinessRequirement/requirements.txt
streamlit
python-dotenv
google-generativeai
ipython
markdown
pdfkit